wall_e_leveling_website/requirements.txt
asgiref==3.7.2
Django==3.2.20
sqlparse==0.4.4
typing_extensions==4.7.1


# needed for the APIs
djangorestframework~=3.14.0
markdown
django-filter

# needed for interacting with db
psycopg2-binary

wall-e-models==0.10
python-dateutil